% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dang_cck81}
\alias{dang_cck81}
\title{Sample table of read counts from the Avana dataset.}
\format{A data frame with 74687 rows (one per sgRNA) and the following 10 columns:
\itemize{
  \item sgRNA: name of the sgRNAs
  \item pDNA_0: plasmid of batch 0 (initial condition of DAN-G).
  \item pDNA_2: plasmid of batch 2 (initial condition of CCK-81).
  \item DAN-G-311Cas9 Rep A p6: replicate A of 1st condition.
  \item DAN-G-311Cas9 Rep B p6: replicate B of 1st condition.
  \item DAN-G-311Cas9 Rep C p6: replicate C of 1st condition.
  \item DAN-G-311Cas9 Rep D p6: replicate D of 1st condition.
  \item CCK-81-311cas9 Rep A p2: replicate A of 2nd condition.
  \item CCK-81-311cas9 Rep B p2: replicate B of 2nd condition.
  \item CCK-81-311cas9 Rep C p2: replicate C of 2nd condition.
}}
\source{
https://portals.broadinstitute.org/achilles
}
\usage{
dang_cck81
}
\description{
The raw integer read counts of the sgRNAs for two conditions from the Avana dataset.
}
\references{
Meyers, R. M. et al. Computational correction of copy number effect improves specificity of CRISPR-Cas9 essentiality screens in cancer cells. Nat Genet (2017). doi:10.1038/ng.3984
}
\keyword{datasets}
