% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dang_cck81_annotation}
\alias{dang_cck81_annotation}
\title{Annotation for the sample table of read counts from the Avana dataset.}
\format{A data frame with 9 rows (one per sample) and the following 8 columns:
\itemize{
  \item Sample: name of the sample, which is equal to the column name of the dataset.
  \item Zeros: number of sgRNAs with zero counts in the corresponding sample.
  \item Gini: gini index of the corresponding sample.
  \item Cell_line: name of the cell line for the corresponding sample.
  \item Treatment: name of the treatment for the corresponding sample.
  \item Timepoint: name of the timepoint for the corresponding sample.
  \item Replicate: name of the replicate for the corresponding sample.
  \item Batch: name of the batch for the corresponding sample.
}}
\usage{
dang_cck81_annotation
}
\description{
An example of an annotation file for the sample from the Avana dataset.
}
\keyword{datasets}
