% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exits}
\alias{exits}
\title{Exit information for Mopac.}
\format{
A tibble with 32 rows and 4 variables:
\describe{
  \item{mi}{distance from Loop 1 southern terminus, in miles}
  \item{km}{distance from Loop 1 southern terminus, in kilometers}
  \item{exit}{main road accessed by exit}
  \item{notes}{additional information}
}
}
\source{
Scraped from \url{https://en.wikipedia.org/wiki/Texas_State_Highway_Loop_1}
}
\usage{
exits
}
\description{
Accurate as of 2021-03-26. Provides basic details for all exits along Loop 1 "Mopac" in Austin, Texas. Exits are unnumbered.
}
\keyword{datasets}
