% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{express}
\alias{express}
\title{Mopac Express Lane Records}
\format{
A tibble with 13032 rows and 6 variables:
\describe{
  \item{direction}{\strong{north}: Far West entry point heading to Parmer, \strong{south}: RM 2222 heading to Cesar Chavez}
  \item{time}{time of day}
  \item{plate}{vehicle license plate}
  \item{make}{vehicle make}
  \item{model}{vehicle model}
  \item{color}{vehicle color}
}
}
\source{
To see how this dataset was generated, visit \url{https://sccm.io/post/mopac-dataset/}

For City of Austin Traffic Counts, visit \url{https://data.austintexas.gov/Transportation-and-Mobility/Camera-Traffic-Counts/sh59-i6y9}
}
\usage{
express
}
\description{
Simulated observations of vehicles entering Mopac express lane at RM 2222 / Far West Blvd. Vehicle make/model/color frequencies are based on actual data from rush_hour. Traffic density inferred from ad hoc sampling and City of Austin Traffic Counts.
}
\keyword{datasets}
