% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonessentials}
\alias{nonessentials}
\title{Hart non-essential genes used as negative controls.}
\format{A character vector}
\source{
https://www.ncbi.nlm.nih.gov/pubmed/24987113
}
\usage{
nonessentials
}
\description{
The names of the non-essential genes used for categorization.
}
\references{
Hart, T., Brown, K. R., Sircoulomb, F., Rottapel, R. & Moffat, J. Measuring error rates in genomic perturbation screens: gold standards for human functional genomics. Mol. Syst. Biol. 10, (2014).
}
\keyword{datasets}
