% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rates}
\alias{rates}
\title{Mopac Express Lane Rates}
\format{
A tibble with 25 rows and 3 variables:
\describe{
  \item{direction}{\strong{north}: Far West entry point heading to Parmer, \strong{south}: RM 2222 heading to Cesar Chavez}
  \item{time}{time of day}
  \item{rate}{TxTag rate in USD}
}
}
\source{
{Rates obtained from Central Texas Regional Mobility Authority.} \url{https://www.mobilityauthority.com/pay-your-toll/rates}
}
\usage{
rates
}
\description{
Express Lane Rates for Loop 1 "Mopac" in Austin, Texas. Rates correspond to RM 2222 / Far West entry points with pre-COVID weekday traffic. Recommend loading \href{https://readr.tidyverse.org/}{readr} package so that \code{time} displays correctly as \strong{hms}.
}
\keyword{datasets}
