% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfastq.R
\name{read.fastq}
\alias{read.fastq}
\title{Map FASTQ files to an sgRNA library.}
\usage{
read.fastq(id = "out", library = NULL, fastq = NULL,
  spacer_start = 0, spacer_length = 19, reverse = TRUE,
  complement = TRUE, essential = NULL, nonessential = NULL,
  nontargeting = NULL, shinyF = NULL)
}
\arguments{
\item{id}{working directory. All future output will be stored in a directory with this name. Defaults to "out".}

\item{library}{sgRNA library. It may be a data frame or a file path. It must contain a column called \code{"sgRNA"} and another called \code{"Gene"}. It may also contain a \code{"Category"} column. Only necessary if the table of counts does not include a column called "Gene".}

\item{fastq}{path of the directory where the FASTQ files are located. All files should have .fastq extension.}

\item{spacer_start}{integer value (starting from 0) specifying the starting position of the sgRNA sequences. Defaults to \code{0}.}

\item{spacer_length}{integer value specifying the length of the sgRNA sequences. Defaults to \code{19}.}

\item{reverse}{logical. Should sgRNA sequences be reversed before mapping? Defaults to \code{TRUE}.}

\item{complement}{logical. Should sgRNA sequences be complemented before mapping? Defaults to \code{TRUE}.}

\item{essential}{an optional vector of gene names to add genes categorized as essential.}

\item{nonessential}{an optional vector of gene names to add genes categorized as non-essential.}

\item{nontargeting}{an optional vector of "gene" names to add the default set of "genes" categorized as non-targeting.}

\item{shinyF}{logical. An internal parameter used by the graphical user interface. If using command-line mode leave as \code{NULL}.}
}
\value{
A list containing the following components:
\item{Library}{The sgRNA library.}
\item{Annotation}{The table of fastq file annotation.}
\item{Counts}{The table of read counts at the FASTQ file level.}
}
\description{
Maps the annotation in all FASTQ files contained in a specified directory to a custom sgRNA library and generates a sample annotation file which needs to be filled out for further analysis.
}
\details{
The following files are produced in the working directory:
\itemize{
\item{\code{fastq_library.txt}, containing two columns: \code{"sgRNA"} and \code{"Gene"} for each sgRNA mapped to the library file, and possibly a \code{"Category"} column as well if supplied.}
\item{\code{fastq_annotation.csv} and \code{fastq_annotation.xlsx}, two annotation files, either of which may be chosen to fill out the following FASTQ file information: cell line, treatment, timepoint, replicate and batch. The annotation files also specify the proportion and percentage of annotation mapped, number of zero counts and gini index per FASTQ file.}
\item{\code{fastq_counts.txt}, containing the read counts per sgRNA for each FASTQ file.}
\item{\code{fastq_reads.txt}, containing the number of reads per FASTQ file.}
}
}
\examples{
## Examples and sample FASTQ files available at:

## reads <- read.fastq(id="out",fastq="dir",library="lib.txt",
##                     spacer_start=0,spacer_length=19,reverse=TRUE,complement=TRUE)

##--------------------------------------------------------------------------------------

}
\author{
Oscar D Villarreal, \email{oscardvillarreal@gmail.com}
}
\keyword{fastq}
\keyword{read}
