% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rra1tail.R
\name{RRA.1tail}
\alias{RRA.1tail}
\title{One-tail robust rank aggregation}
\usage{
RRA.1tail(level1 = NULL, level2 = NULL, score = NULL)
}
\arguments{
\item{level1}{Vector containing genes.}

\item{level2}{Vector containing sgRNAs.}

\item{score}{Vector containing sgRNA scores.}
}
\value{
A data frame with modified robust-rank-aggregation p-value.
}
\description{
Performs one-tail a-RRA analysis using the original algorithm of MAGeCK-RRA.
}
\references{
Li, W. et al. MAGeCK enables robust identification of essential genes from genome-scale CRISPR/Cas9 knockout screens. Genome Biol. 15, 554 (2014).

##--------------------------------------------------------------------------------------
}
\author{
Oscar D Villarreal, \email{oscardvillarreal@gmail.com}
}
\keyword{RRA}
