% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rra2tail.R
\name{RRA.2tail}
\alias{RRA.2tail}
\title{Gene essentiality 2-tail robust rank aggregation}
\usage{
RRA.2tail(id = NULL, sgRNA = NULL, conditions = NULL,
  pvalue = 0.05, fraction = 0.8, shinyF = NULL)
}
\arguments{
\item{id}{an optional identifier for the analysis. All input will be read from and all output will be stored in a directory with this name. If \code{NULL}, two data frames must be supplied containing the sgRNA-level and gene-level replicate-averaged log-fold-changes. Defaults to \code{NULL}.}

\item{sgRNA}{an optional data frame containing the replicate-averaged log-fold-changes for each sgRNA. If \code{NULL}, the corresponding file in the identifier directory will be read.}

\item{conditions}{a vector of condition names to be analyzed. All conditions must be column names in the replicate-averaged log-fold-change datasets. Defaults to all conditions in the dataset.}

\item{pvalue}{numeric. P-value threshold for depletion/enrichment. Defaults to 0.05.}

\item{fraction}{numeric. Mininmum fraction of conditions in which a gene must satisfy the p-value threshold. Defaults to 0.8 (i.e. the gene should be significant in at least 80\% of all conditions).}

\item{shinyF}{logical. An internal parameter used by the graphical user interface. If using command-line leave as \code{NULL}.}
}
\value{
A list containing the following components:
\item{p.value}{List of tables of p values and FDR for all genes of each condition, indicating the ranking of depletion/enrichment.}
\item{Essential}{List of filtered essential genes (i.e. those whose 2-tail p value is smaller than the threshold in the minimal fraction of conditions).}
\item{Nonessential}{List of filtered non-essential genes (i.e. those whose 2-tail p value is larger than the threshold in the minimal fraction of conditions).}
\item{Nondifferential}{List of genes which were not found to be significantly differentially essential.}
\item{Enriched}{List of genes which were found to be significantly enriched.}
\item{Unenriched}{List of genes which were not found to be significantly enriched in any of the conditions analyzed.}
}
\description{
Performs 2-tail a-RRA analysis of gene essentiality. Within MoPAC's framework, this is necessary if there are no control genes available for the normalization of two conditions when computing the differential essentiality scores, or if the control genes should be filtered when computing the essentiality scores.
}
\details{
A modified version of the alpha-robust-rank-aggregation algorithm from MAGeCK-RRA is implemented whereby the skewness of the sgRNA distributions is simultaneously assessed for both negative and positive selection.
A gene with a significant p value may thus be either significantly depleted or significantly enriched in the screen, and it is therefore necessary to consider the sign of its log-fold-change as well in order to distinguish between negative and positive selection.

If \code{id} is not \code{NULL}, the following files are produced in the directory with name equal to the identifier:
\itemize{
 \item{\code{RRA.txt} files, one for each condition.}
 \item{\code{essential.txt}, \code{nonessential.txt}, \code{nondifferential.txt}, \code{enriched.txt} and \code{unenriched.txt}, the filtered lists of essential, nonessential, nondifferential, enriched and unenriched genes.}
}
}
\examples{

## Example I: Two-tail RRA using an identifier to produce output:

## 1. Prepare read counts information:
counts <- read.counts(id="DANG_CCK81", counts=MoPAC::dang_cck81, library=MoPAC::sgRNA_library)

## 2. Manually fill out either of the two annotation files produced by read.counts.
## In this example we will fill it out using R:
annotation <- counts$Annotation
annotation$Condition <- c("Plasmid0","Plasmid1",rep("DANG",4),rep("CCK81",3))
annotation$Replicate <- c("A","A","A","B","C","D","A","B","C")
annotation$Control <- c("","",rep("Plasmid0",4),rep("Plasmid1",3))
utils::write.csv(annotation,"DANG_CCK81/annotation.csv",quote=FALSE,row.names=FALSE)

## 3. Get pre-processed fold changes:
qc <- quality.control(id="DANG_CCK81")

## 4. Finally get the two-tail RRA essentiality analysis:
significance <- RRA.2tail(id="DANG_CCK81")

## Example II: Two-tail RRA without using an identifier:

## 1. Prepare read counts information:
counts <- read.counts(counts=MoPAC::dang_cck81, library=MoPAC::sgRNA_library)

## 2. Prepare sample annotation as a data frame:
annotation <- counts$Annotation
annotation$Condition <- c("Plasmid0","Plasmid1",rep("DANG",4),rep("CCK81",3))
annotation$Replicate <- c("A","A","A","B","C","D","A","B","C")
annotation$Control <- c("","",rep("Plasmid0",4),rep("Plasmid1",3))

## 3. Get pre-processed fold changes:
qc <- quality.control(counts=counts$Counts, annotation=annotation)

## 4. Finally get the two-tail RRA essentiality analysis:
significance <- RRA.2tail(sgRNA=qc$sgRNA_lfc)

##--------------------------------------------------------------------------------------

}
\references{
Li, W. et al. MAGeCK enables robust identification of essential genes from genome-scale CRISPR/Cas9 knockout screens. Genome Biol. 15, 554 (2014).
}
\author{
Oscar D Villarreal, \email{oscardvillarreal@gmail.com}
}
\keyword{RRA}
\keyword{essentiality}
