% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{run.MoPAC}
\alias{run.MoPAC}
\title{MoPAC pipeline}
\usage{
run.MoPAC()
}
\description{
Interactive use of the MoPAC pipeline.
}
\author{
Oscar D Villarreal, \email{oscardvillarreal@gmail.com}
}
\keyword{gui}
\keyword{mopac}
\keyword{pipeline}
\keyword{shiny}
