% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sgRNA_library}
\alias{sgRNA_library}
\title{Sample sgRNA library.}
\format{A data frame containing 74768 rows (one per sgRNA) and the following 3 columns:
\itemize{
  \item sgRNA: name of the sgRNAs.
  \item Gene: gene names for each sgRNA.
  \item ID: Hugo IDs for each gene.
}}
\source{
https://portals.broadinstitute.org/achilles
}
\usage{
sgRNA_library
}
\description{
A sample sgRNA library.
}
\references{
Aguirre, A. J. et al. Genomic Copy Number Dictates a Gene-Independent Cell Response to CRISPR/Cas9 Targeting. Cancer Discov. 6, 914–929 (2016).
}
\keyword{datasets}
