% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stringdb_9606}
\alias{stringdb_9606}
\title{STRINGdb datasets}
\format{A list of data frames containing protein interactions in homo sapiens downloaded from the STRING consortium}
\source{
http://www.string-db.org
}
\usage{
stringdb_9606
}
\description{
STRINGdb datasets
}
\references{
Szklarczyk, D. et al. The STRING database in 2017: Quality-controlled protein-protein association networks, made broadly accessible. Nucleic Acids Res. 45, D362–D368 (2017).
}
\keyword{datasets}
